refactor(controller): split proxy code into deployment and provider files
- Extract gateway deployment functions from claw_proxy.go into claw_deployment.go
  (configureClawDeploymentForVertex, configureClawDeploymentForKubernetes,
  configureClawDeploymentConfigMode, stampGatewayConfigHash, etc.)
- Extract shared provider utilities into claw_providers.go
  (resolveProviderDefaults, resolveProviderInfo, usesVertexSDK, etc.)
- Reorganize tests to match new file structure: create claw_providers_test.go,
  move tests between files so each test lives alongside its source
- claw_proxy.go now contains only proxy-specific logic
