+++ internal/controller/claw_credentials_test.go
@@ -18,14 +18,12 @@ package controller
 
 import (
 	"context"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -324,230 +322,6 @@ func TestOpenClawCredentialSecretReference(t *testing.T) {
 
 }
 
-func TestConfigureProxyForCredentials(t *testing.T) {
-	buildObjects := func(t *testing.T) (*clawv1alpha1.Claw, []*unstructured.Unstructured) {
-		t.Helper()
-		reconciler := createClawReconciler()
-		instance := &clawv1alpha1.Claw{}
-		instance.Name = testInstanceName
-		instance.Namespace = namespace
-		objects, err := reconciler.buildKustomizedObjects(instance)
-		require.NoError(t, err)
-		return instance, objects
-	}
-
-	findProxyContainer := func(t *testing.T, objects []*unstructured.Unstructured) map[string]any {
-		t.Helper()
-		for _, obj := range objects {
-			if obj.GetKind() == DeploymentKind && obj.GetName() == getProxyDeploymentName(testInstanceName) {
-				containers, _, _ := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "containers")
-				require.NotEmpty(t, containers)
-				c, ok := containers[0].(map[string]any)
-				require.True(t, ok)
-				return c
-			}
-		}
-		t.Fatal("proxy deployment not found")
-		return nil
-	}
-
-	findVolumes := func(t *testing.T, objects []*unstructured.Unstructured) []any {
-		t.Helper()
-		for _, obj := range objects {
-			if obj.GetKind() == DeploymentKind && obj.GetName() == getProxyDeploymentName(testInstanceName) {
-				vols, _, _ := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "volumes")
-				return vols
-			}
-		}
-		return nil
-	}
-
-	t.Run("should add GCP volume and mount for gcp credential", func(t *testing.T) {
-		instance, objects := buildObjects(t)
-		creds := []clawv1alpha1.CredentialSpec{
-			{
-				Name:      "vertex",
-				Type:      clawv1alpha1.CredentialTypeGCP,
-				SecretRef: &clawv1alpha1.SecretRef{Name: "gcp-sa", Key: "sa.json"},
-				Domain:    ".googleapis.com",
-				GCP:       &clawv1alpha1.GCPConfig{Project: "p", Location: "us-central1"},
-			},
-		}
-		require.NoError(t, configureProxyForCredentials(objects, instance, toResolved(creds)))
-
-		container := findProxyContainer(t, objects)
-		mounts, _, _ := unstructured.NestedSlice(container, "volumeMounts")
-
-		var foundMount bool
-		for _, m := range mounts {
-			mount := m.(map[string]any)
-			if mount["name"] == "cred-vertex" {
-				assert.Equal(t, "/etc/proxy/credentials/vertex", mount["mountPath"])
-				assert.Equal(t, true, mount["readOnly"])
-				foundMount = true
-			}
-		}
-		assert.True(t, foundMount, "GCP credential volume mount should be present")
-
-		volumes := findVolumes(t, objects)
-		var foundVol bool
-		for _, v := range volumes {
-			vol := v.(map[string]any)
-			if vol["name"] == "cred-vertex" {
-				foundVol = true
-				secret, ok := vol["secret"].(map[string]any)
-				require.True(t, ok)
-				assert.Equal(t, "gcp-sa", secret["secretName"])
-			}
-		}
-		assert.True(t, foundVol, "GCP credential volume should be present")
-	})
-
-	t.Run("should skip credentials with nil secretRef for apiKey type", func(t *testing.T) {
-		instance, objects := buildObjects(t)
-		creds := []clawv1alpha1.CredentialSpec{
-			{
-				Name:   "no-ref",
-				Type:   clawv1alpha1.CredentialTypeAPIKey,
-				Domain: "api.example.com",
-				APIKey: &clawv1alpha1.APIKeyConfig{Header: "x-api-key"},
-			},
-		}
-		require.NoError(t, configureProxyForCredentials(objects, instance, toResolved(creds)))
-
-		container := findProxyContainer(t, objects)
-		envVars, _, _ := unstructured.NestedSlice(container, "env")
-		for _, e := range envVars {
-			env := e.(map[string]any)
-			assert.NotEqual(t, "CRED_NO_REF", env["name"], "should not add env var for credential without secretRef")
-		}
-	})
-
-	t.Run("should handle multiple credential types together", func(t *testing.T) {
-		instance, objects := buildObjects(t)
-		creds := []clawv1alpha1.CredentialSpec{
-			{
-				Name:      "gemini",
-				Type:      clawv1alpha1.CredentialTypeAPIKey,
-				SecretRef: &clawv1alpha1.SecretRef{Name: "s1", Key: "k1"},
-				Domain:    ".googleapis.com",
-				APIKey:    &clawv1alpha1.APIKeyConfig{Header: "x-goog-api-key"},
-			},
-			{
-				Name:      "openai",
-				Type:      clawv1alpha1.CredentialTypeBearer,
-				SecretRef: &clawv1alpha1.SecretRef{Name: "s2", Key: "k2"},
-				Domain:    "api.openai.com",
-			},
-		}
-		require.NoError(t, configureProxyForCredentials(objects, instance, toResolved(creds)))
-
-		container := findProxyContainer(t, objects)
-		envVars, _, _ := unstructured.NestedSlice(container, "env")
-
-		envNames := make(map[string]bool)
-		for _, e := range envVars {
-			env := e.(map[string]any)
-			envNames[env["name"].(string)] = true
-		}
-		assert.True(t, envNames["CRED_GEMINI"], "should have CRED_GEMINI")
-		assert.True(t, envNames["CRED_OPENAI"], "should have CRED_OPENAI")
-	})
-}
-
-func TestStampSecretVersionAnnotation(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run("should stamp Secret ResourceVersion on proxy pod template", func(t *testing.T) {
-		t.Cleanup(func() { deleteAndWaitAllResources(t, namespace) })
-
-		createClawInstance(t, ctx, testInstanceName, namespace)
-		reconciler := createClawReconciler()
-		reconcileClaw(t, ctx, reconciler, testInstanceName, namespace)
-
-		deployment := &appsv1.Deployment{}
-		require.NoError(t, k8sClient.Get(ctx, client.ObjectKey{
-			Name:      getProxyDeploymentName(testInstanceName),
-			Namespace: namespace,
-		}, deployment))
-
-		annotations := deployment.Spec.Template.Annotations
-		require.NotNil(t, annotations, "pod template annotations should exist")
-		geminiSecretVersionKey := clawv1alpha1.AnnotationPrefixSecretVersion + "gemini" + clawv1alpha1.AnnotationSuffixSecretVersion
-		rv, ok := annotations[geminiSecretVersionKey]
-		assert.True(t, ok, "gemini-secret-version annotation should exist")
-		assert.NotEmpty(t, rv, "ResourceVersion should not be empty")
-	})
-
-	t.Run("should update annotation when Secret data changes", func(t *testing.T) {
-		t.Cleanup(func() { deleteAndWaitAllResources(t, namespace) })
-
-		createClawInstance(t, ctx, testInstanceName, namespace)
-		reconciler := createClawReconciler()
-		reconcileClaw(t, ctx, reconciler, testInstanceName, namespace)
-
-		deployment := &appsv1.Deployment{}
-		require.NoError(t, k8sClient.Get(ctx, client.ObjectKey{
-			Name:      getProxyDeploymentName(testInstanceName),
-			Namespace: namespace,
-		}, deployment))
-		geminiSecretVersionKey := clawv1alpha1.AnnotationPrefixSecretVersion + "gemini" + clawv1alpha1.AnnotationSuffixSecretVersion
-		originalRV := deployment.Spec.Template.Annotations[geminiSecretVersionKey]
-		require.NotEmpty(t, originalRV)
-
-		// Update the Secret data
-		secret := &corev1.Secret{}
-		require.NoError(t, k8sClient.Get(ctx, client.ObjectKey{
-			Name:      aiModelSecret,
-			Namespace: namespace,
-		}, secret))
-		secret.Data[aiModelSecretKey] = []byte("rotated-api-key")
-		require.NoError(t, k8sClient.Update(ctx, secret))
-
-		// Reconcile again
-		reconcileClaw(t, ctx, reconciler, testInstanceName, namespace)
-
-		require.NoError(t, k8sClient.Get(ctx, client.ObjectKey{
-			Name:      getProxyDeploymentName(testInstanceName),
-			Namespace: namespace,
-		}, deployment))
-		newRV := deployment.Spec.Template.Annotations[geminiSecretVersionKey]
-		assert.NotEqual(t, originalRV, newRV,
-			"annotation should change after Secret data update (old=%s, new=%s)", originalRV, newRV)
-	})
-
-	t.Run("should skip credentials without secretRef", func(t *testing.T) {
-		t.Cleanup(func() { deleteAndWaitAllResources(t, namespace) })
-
-		instance := &clawv1alpha1.Claw{}
-		instance.Name = testInstanceName
-		instance.Namespace = namespace
-		instance.Spec.Credentials = []clawv1alpha1.CredentialSpec{
-			{
-				Name:   "passthrough",
-				Type:   clawv1alpha1.CredentialTypeNone,
-				Domain: "example.com",
-			},
-		}
-		require.NoError(t, k8sClient.Create(ctx, instance))
-
-		reconciler := createClawReconciler()
-		reconcileClaw(t, ctx, reconciler, testInstanceName, namespace)
-
-		deployment := &appsv1.Deployment{}
-		require.NoError(t, k8sClient.Get(ctx, client.ObjectKey{
-			Name:      getProxyDeploymentName(testInstanceName),
-			Namespace: namespace,
-		}, deployment))
-
-		annotations := deployment.Spec.Template.Annotations
-		for key := range annotations {
-			assert.False(t, strings.HasSuffix(key, clawv1alpha1.AnnotationSuffixSecretVersion),
-				"should not have secret-version annotations for none-type credentials, found %s", key)
-		}
-	})
-}
-
 func TestFindClawsReferencingSecret(t *testing.T) {
 	ctx := context.Background()
 
@@ -591,149 +365,4 @@ func TestFindClawsReferencingSecret(t *testing.T) {
 		requests := reconciler.findClawsReferencingSecret(ctx, secret)
 		assert.Empty(t, requests)
 	})
-}
-
-// --- Provider defaults tests ---
-
-func TestResolveProviderDefaults(t *testing.T) {
-	tests := []struct {
-		name       string
-		cred       clawv1alpha1.CredentialSpec
-		wantDomain string
-		wantHeader string
-		wantErr    string
-	}{
-		{
-			name: "google apiKey fills domain and header",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-			},
-			wantDomain: "generativelanguage.googleapis.com",
-			wantHeader: "x-goog-api-key",
-		},
-		{
-			name: "anthropic apiKey fills domain and header",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "anthropic",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "anthropic",
-			},
-			wantDomain: "api.anthropic.com",
-			wantHeader: "x-api-key",
-		},
-		{
-			name: "google gcp fills domain",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "google",
-				GCP:      &clawv1alpha1.GCPConfig{Project: "p", Location: "us-central1"},
-			},
-			wantDomain: ".googleapis.com",
-		},
-		{
-			name: "anthropic gcp fills domain",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "anthropic-vertex",
-				Type:     clawv1alpha1.CredentialTypeGCP,
-				Provider: "anthropic",
-				GCP:      &clawv1alpha1.GCPConfig{Project: "p", Location: "us-east5"},
-			},
-			wantDomain: ".googleapis.com",
-		},
-		{
-			name: "explicit domain preserved",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-				Domain:   "custom-proxy.internal",
-			},
-			wantDomain: "custom-proxy.internal",
-			wantHeader: "x-goog-api-key",
-		},
-		{
-			name: "explicit apiKey preserved",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "gemini",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "google",
-				APIKey:   &clawv1alpha1.APIKeyConfig{Header: "x-custom-key"},
-			},
-			wantDomain: "generativelanguage.googleapis.com",
-			wantHeader: "x-custom-key",
-		},
-		{
-			name: "unknown provider with domain succeeds",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "custom-llm",
-				Domain:   "api.custom-llm.com",
-				APIKey:   &clawv1alpha1.APIKeyConfig{Header: "x-api-key"},
-			},
-			wantDomain: "api.custom-llm.com",
-			wantHeader: "x-api-key",
-		},
-		{
-			name: "unknown provider without domain errors",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "custom-llm",
-				APIKey:   &clawv1alpha1.APIKeyConfig{Header: "x-api-key"},
-			},
-			wantErr: "domain is required",
-		},
-		{
-			name: "unknown provider without apiKey errors",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeAPIKey,
-				Provider: "custom-llm",
-				Domain:   "api.custom-llm.com",
-			},
-			wantErr: "apiKey config is required",
-		},
-		{
-			name: "no provider with domain and apiKey succeeds",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:   "legacy",
-				Type:   clawv1alpha1.CredentialTypeAPIKey,
-				Domain: "api.example.com",
-				APIKey: &clawv1alpha1.APIKeyConfig{Header: "x-token"},
-			},
-			wantDomain: "api.example.com",
-			wantHeader: "x-token",
-		},
-		{
-			name: "bearer type with no domain errors",
-			cred: clawv1alpha1.CredentialSpec{
-				Name:     "custom",
-				Type:     clawv1alpha1.CredentialTypeBearer,
-				Provider: "custom-llm",
-			},
-			wantErr: "domain is required",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			cred := tt.cred
-			err := resolveProviderDefaults(&cred)
-			if tt.wantErr != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.wantErr)
-				return
-			}
-			require.NoError(t, err)
-			assert.Equal(t, tt.wantDomain, cred.Domain)
-			if tt.wantHeader != "" {
-				require.NotNil(t, cred.APIKey)
-				assert.Equal(t, tt.wantHeader, cred.APIKey.Header)
-			}
-		})
-	}
 }
