+++ internal/controller/claw_kubernetes_test.go
@@ -18,7 +18,6 @@ package controller
 
 import (
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"strings"
 	"testing"
@@ -321,246 +320,6 @@ func TestSanitizeKubeconfig(t *testing.T) {
 	})
 }
 
-// --- generateProxyConfig kubernetes route tests ---
-
-func TestGenerateProxyConfigKubernetes(t *testing.T) {
-	t.Run("should generate routes per cluster from kubeconfig", func(t *testing.T) {
-		creds := []resolvedCredential{
-			{
-				CredentialSpec: clawv1alpha1.CredentialSpec{
-					Name:      "k8s",
-					Type:      clawv1alpha1.CredentialTypeKubernetes,
-					SecretRef: &clawv1alpha1.SecretRef{Name: testKubeSecretName, Key: "config"},
-				},
-				KubeConfig: &kubeconfigData{
-					Clusters: []kubeconfigCluster{
-						{Name: "prod", Hostname: "api.prod.example.com", Port: "6443"},
-						{Name: "staging", Hostname: "api.staging.example.com", Port: "443"},
-					},
-				},
-			},
-		}
-
-		data, err := generateProxyConfig(creds)
-		require.NoError(t, err)
-
-		var cfg proxyConfig
-		require.NoError(t, json.Unmarshal(data, &cfg))
-
-		prodRoute := findRouteByDomain(t, cfg.Routes, "api.prod.example.com:6443")
-		assert.Equal(t, "kubernetes", prodRoute.Injector)
-		assert.Equal(t, "/etc/proxy/credentials/k8s/kubeconfig", prodRoute.KubeconfigPath)
-
-		stagingRoute := findRouteByDomain(t, cfg.Routes, "api.staging.example.com:443")
-		assert.Equal(t, "kubernetes", stagingRoute.Injector)
-	})
-
-	t.Run("should skip kubernetes credential with nil kubeconfig", func(t *testing.T) {
-		creds := []resolvedCredential{
-			{
-				CredentialSpec: clawv1alpha1.CredentialSpec{
-					Name: "k8s",
-					Type: clawv1alpha1.CredentialTypeKubernetes,
-				},
-				KubeConfig: nil,
-			},
-		}
-
-		data, err := generateProxyConfig(creds)
-		require.NoError(t, err)
-
-		var cfg proxyConfig
-		require.NoError(t, json.Unmarshal(data, &cfg))
-		// Only builtin passthrough routes
-		assert.Len(t, cfg.Routes, len(builtinPassthroughDomains))
-	})
-
-	t.Run("should include caCert when cluster has CAData", func(t *testing.T) {
-		caData := []byte("-----BEGIN CERTIFICATE-----\nMIIBfake\n-----END CERTIFICATE-----\n")
-		creds := []resolvedCredential{
-			{
-				CredentialSpec: clawv1alpha1.CredentialSpec{
-					Name:      "k8s",
-					Type:      clawv1alpha1.CredentialTypeKubernetes,
-					SecretRef: &clawv1alpha1.SecretRef{Name: testKubeSecretName, Key: "config"},
-				},
-				KubeConfig: &kubeconfigData{
-					Clusters: []kubeconfigCluster{
-						{Name: "prod", Hostname: "api.prod.example.com", Port: "6443", CAData: caData},
-						{Name: "staging", Hostname: "api.staging.example.com", Port: "443"},
-					},
-				},
-			},
-		}
-
-		data, err := generateProxyConfig(creds)
-		require.NoError(t, err)
-
-		var cfg proxyConfig
-		require.NoError(t, json.Unmarshal(data, &cfg))
-
-		prodRoute := findRouteByDomain(t, cfg.Routes, "api.prod.example.com:6443")
-		assert.NotEmpty(t, prodRoute.CACert, "route with CAData should have caCert set")
-
-		decoded, err := base64.StdEncoding.DecodeString(prodRoute.CACert)
-		require.NoError(t, err)
-		assert.Equal(t, caData, decoded)
-
-		stagingRoute := findRouteByDomain(t, cfg.Routes, "api.staging.example.com:443")
-		assert.Empty(t, stagingRoute.CACert, "route without CAData should have empty caCert")
-	})
-}
-
-// --- configureProxyForCredentials kubernetes volume mount tests ---
-
-func TestConfigureProxyForKubernetesCredentials(t *testing.T) {
-	buildObjects := func(t *testing.T) (*clawv1alpha1.Claw, []*unstructured.Unstructured) {
-		t.Helper()
-		reconciler := createClawReconciler()
-		instance := &clawv1alpha1.Claw{}
-		instance.Name = testInstanceName
-		instance.Namespace = namespace
-		objects, err := reconciler.buildKustomizedObjects(instance)
-		require.NoError(t, err)
-		return instance, objects
-	}
-
-	t.Run("should add kubernetes kubeconfig volume mount", func(t *testing.T) {
-		instance, objects := buildObjects(t)
-		creds := []resolvedCredential{
-			{
-				CredentialSpec: clawv1alpha1.CredentialSpec{
-					Name:      "k8s",
-					Type:      clawv1alpha1.CredentialTypeKubernetes,
-					SecretRef: &clawv1alpha1.SecretRef{Name: testKubeSecretName, Key: "config"},
-				},
-			},
-		}
-
-		require.NoError(t, configureProxyForCredentials(objects, instance, creds))
-
-		for _, obj := range objects {
-			if obj.GetKind() != DeploymentKind || obj.GetName() != getProxyDeploymentName(testInstanceName) {
-				continue
-			}
-			volumes, _, _ := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "volumes")
-			var foundVol bool
-			for _, v := range volumes {
-				vol := v.(map[string]any)
-				if vol["name"] == testKubeCredVolume {
-					foundVol = true
-					secret := vol["secret"].(map[string]any)
-					assert.Equal(t, testKubeSecretName, secret["secretName"])
-					items := secret["items"].([]any)
-					item := items[0].(map[string]any)
-					assert.Equal(t, "config", item["key"])
-					assert.Equal(t, "kubeconfig", item["path"])
-				}
-			}
-			assert.True(t, foundVol, "kubernetes credential volume should be present")
-		}
-	})
-}
-
-// --- configureClawDeploymentForKubernetes tests ---
-
-func TestConfigureClawDeploymentForKubernetes(t *testing.T) {
-	makeDeployment := func() []*unstructured.Unstructured {
-		dep := &unstructured.Unstructured{}
-		dep.SetKind(DeploymentKind)
-		dep.SetName(getClawDeploymentName(testInstanceName))
-		dep.Object["spec"] = map[string]any{
-			"template": map[string]any{
-				"spec": map[string]any{
-					"containers": []any{
-						map[string]any{
-							"name":         ClawGatewayContainerName,
-							"env":          []any{},
-							"volumeMounts": []any{},
-						},
-					},
-					"volumes": []any{},
-				},
-			},
-		}
-		return []*unstructured.Unstructured{dep}
-	}
-
-	t.Run("should add KUBECONFIG env, PATH, volumes, and init container", func(t *testing.T) {
-		objects := makeDeployment()
-		creds := []resolvedCredential{
-			{
-				CredentialSpec: clawv1alpha1.CredentialSpec{
-					Name:      "k8s",
-					Type:      clawv1alpha1.CredentialTypeKubernetes,
-					SecretRef: &clawv1alpha1.SecretRef{Name: testKubeSecretName, Key: "config"},
-				},
-				KubeConfig: &kubeconfigData{},
-			},
-		}
-
-		require.NoError(t, configureClawDeploymentForKubernetes(objects, creds, DefaultKubectlImage, testInstanceName))
-
-		containers, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "containers")
-		container := containers[0].(map[string]any)
-
-		envVars := container["env"].([]any)
-		envMap := map[string]string{}
-		for _, e := range envVars {
-			env := e.(map[string]any)
-			envMap[env["name"].(string)] = env["value"].(string)
-		}
-		assert.Equal(t, "/etc/kube/config", envMap["KUBECONFIG"])
-		assert.Contains(t, envMap["PATH"], "/opt/kube-tools")
-
-		volumeMounts := container["volumeMounts"].([]any)
-		require.Len(t, volumeMounts, 2)
-		vmNames := map[string]string{}
-		for _, vm := range volumeMounts {
-			m := vm.(map[string]any)
-			vmNames[m["name"].(string)] = m["mountPath"].(string)
-		}
-		assert.Equal(t, "/etc/kube", vmNames["kube-config"])
-		assert.Equal(t, "/opt/kube-tools", vmNames["kubectl-bin"])
-
-		volumes, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "volumes")
-		require.Len(t, volumes, 2)
-		volNames := map[string]bool{}
-		for _, v := range volumes {
-			vol := v.(map[string]any)
-			volNames[vol["name"].(string)] = true
-		}
-		assert.True(t, volNames["kube-config"])
-		assert.True(t, volNames["kubectl-bin"])
-
-		initContainers, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "initContainers")
-		require.Len(t, initContainers, 1)
-		initC := initContainers[0].(map[string]any)
-		assert.Equal(t, "init-kubectl", initC["name"])
-		assert.Equal(t, DefaultKubectlImage, initC["image"])
-	})
-
-	t.Run("should be no-op when no kubernetes credentials exist", func(t *testing.T) {
-		objects := makeDeployment()
-		creds := []resolvedCredential{
-			{
-				CredentialSpec: clawv1alpha1.CredentialSpec{
-					Name:   "gemini",
-					Type:   clawv1alpha1.CredentialTypeAPIKey,
-					Domain: "generativelanguage.googleapis.com",
-				},
-			},
-		}
-
-		require.NoError(t, configureClawDeploymentForKubernetes(objects, creds, DefaultKubectlImage, testInstanceName))
-
-		containers, _, _ := unstructured.NestedSlice(objects[0].Object, "spec", "template", "spec", "containers")
-		container := containers[0].(map[string]any)
-		envVars := container["env"].([]any)
-		assert.Empty(t, envVars)
-	})
-}
-
 // --- injectKubePortsIntoNetworkPolicy tests ---
 
 func TestInjectKubePortsIntoNetworkPolicy(t *testing.T) {
@@ -746,20 +505,6 @@ func TestInjectKubernetesSkill(t *testing.T) {
 
 		_, found, _ := unstructured.NestedString(objects[0].Object, "data", "KUBERNETES.md")
 		assert.False(t, found, "KUBERNETES.md should not exist when no kubernetes credentials")
-	})
-}
-
-// --- resolveProviderDefaults kubernetes tests ---
-
-func TestResolveProviderDefaultsKubernetes(t *testing.T) {
-	t.Run("should return nil for kubernetes type (no domain required)", func(t *testing.T) {
-		cred := clawv1alpha1.CredentialSpec{
-			Name: "k8s",
-			Type: clawv1alpha1.CredentialTypeKubernetes,
-		}
-		err := resolveProviderDefaults(&cred)
-		require.NoError(t, err)
-		assert.Empty(t, cred.Domain)
 	})
 }
