+++ internal/controller/claw_proxy.go
@@ -22,7 +22,6 @@ import (
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
-	"crypto/sha256"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/base64"
@@ -76,103 +75,6 @@ type proxyConfig struct {
 func credEnvVarName(credName string) string {
 	upper := strings.ToUpper(credName)
 	return "CRED_" + strings.ReplaceAll(upper, "-", "_")
-}
-
-// usesVertexSDK returns true when a credential should use the native Vertex AI SDK
-// instead of a gateway proxy route. This applies to non-Google GCP providers (e.g.,
-// Anthropic via Vertex AI), where the provider's SDK format doesn't match Vertex AI's
-// URL structure and the native @anthropic-ai/vertex-sdk handles it correctly.
-func usesVertexSDK(cred clawv1alpha1.CredentialSpec) bool {
-	return cred.Type == clawv1alpha1.CredentialTypeGCP && cred.Provider != "" && cred.Provider != "google"
-}
-
-// vertexAIBaseURL returns the Vertex AI REST base URL for the given location.
-// The "global" location uses a plain hostname without a region prefix.
-func vertexAIBaseURL(location string) string {
-	if location == "global" {
-		return "https://aiplatform.googleapis.com"
-	}
-	return "https://" + location + "-aiplatform.googleapis.com"
-}
-
-// vertexProviderAPIMapping maps provider names to their OpenClaw API identifiers for Vertex AI.
-var vertexProviderAPIMapping = map[string]string{
-	"anthropic": "anthropic-messages",
-}
-
-// apiKeyProviderDefault holds the default domain and header for a known provider using type: apiKey.
-type apiKeyProviderDefault struct {
-	Domain string
-	Header string
-}
-
-// knownAPIKeyProviders maps provider names to their default domain and header.
-var knownAPIKeyProviders = map[string]apiKeyProviderDefault{
-	"google":    {Domain: "generativelanguage.googleapis.com", Header: "x-goog-api-key"},
-	"anthropic": {Domain: "api.anthropic.com", Header: "x-api-key"},
-}
-
-// resolveProviderDefaults fills in missing Domain and APIKey fields for known providers.
-// Explicit values are preserved (escape hatch). Returns an error if required fields
-// are still missing after applying defaults (unknown provider without domain/apiKey).
-func resolveProviderDefaults(cred *clawv1alpha1.CredentialSpec) error {
-	switch cred.Type {
-	case clawv1alpha1.CredentialTypeAPIKey:
-		if defaults, ok := knownAPIKeyProviders[cred.Provider]; ok {
-			if cred.Domain == "" {
-				cred.Domain = defaults.Domain
-			}
-			if cred.APIKey == nil {
-				cred.APIKey = &clawv1alpha1.APIKeyConfig{Header: defaults.Header}
-			}
-		}
-
-	case clawv1alpha1.CredentialTypeGCP:
-		if cred.Domain == "" {
-			cred.Domain = ".googleapis.com"
-		}
-
-	case clawv1alpha1.CredentialTypeKubernetes:
-		// Domains are derived from the kubeconfig, not the domain field
-		return nil
-	}
-
-	if cred.Domain == "" {
-		return fmt.Errorf("credential %q: domain is required (no default for provider %q with type %q)", cred.Name, cred.Provider, cred.Type)
-	}
-	if cred.Type == clawv1alpha1.CredentialTypeAPIKey && cred.APIKey == nil {
-		return fmt.Errorf("credential %q: apiKey config is required (no default for provider %q)", cred.Name, cred.Provider)
-	}
-	return nil
-}
-
-// providerInfo holds the resolved upstream host and base path for a provider's gateway route.
-type providerInfo struct {
-	Upstream string
-	BasePath string
-}
-
-// resolveProviderInfo returns the upstream and base path for a credential's provider.
-// GCP credentials route through Vertex AI with the provider name as the publisher
-// (works for google, anthropic, meta, etc.). Google + apiKey uses the Gemini REST API.
-// All other combos: upstream = domain, basePath = "".
-func resolveProviderInfo(cred clawv1alpha1.CredentialSpec) providerInfo {
-	if cred.Type == clawv1alpha1.CredentialTypeGCP && cred.GCP != nil {
-		return providerInfo{
-			Upstream: vertexAIBaseURL(cred.GCP.Location),
-			BasePath: "/v1/projects/" + cred.GCP.Project + "/locations/" + cred.GCP.Location + "/publishers/" + cred.Provider,
-		}
-	}
-
-	if cred.Provider == "google" {
-		return providerInfo{
-			Upstream: "https://generativelanguage.googleapis.com",
-			BasePath: "/v1beta",
-		}
-	}
-
-	domain := strings.TrimPrefix(cred.Domain, ".")
-	return providerInfo{Upstream: "https://" + domain}
 }
 
 // builtinPassthrough defines a domain the proxy always allows without credential injection.
@@ -330,55 +232,6 @@ func (r *ClawResourceReconciler) applyProxyConfigMap(ctx context.Context, instan
 	return nil
 }
 
-// hasVertexSDKCredentials returns true if any credential uses the native Vertex SDK.
-func hasVertexSDKCredentials(credentials []resolvedCredential) bool {
-	for _, rc := range credentials {
-		if usesVertexSDK(rc.CredentialSpec) {
-			return true
-		}
-	}
-	return false
-}
-
-// applyVertexADCConfigMap creates or updates the stub ADC ConfigMap used by the
-// OpenClaw pod's Vertex SDK to bootstrap GCP token refresh. The stub contains
-// dummy credentials — the MITM proxy replaces tokens with real ones.
-func (r *ClawResourceReconciler) applyVertexADCConfigMap(ctx context.Context, instance *clawv1alpha1.Claw, resolvedCreds []resolvedCredential) error {
-	configMapName := getVertexADCConfigMapName(instance.Name)
-	if !hasVertexSDKCredentials(resolvedCreds) {
-		existing := &corev1.ConfigMap{}
-		if err := r.Get(ctx, client.ObjectKey{Name: configMapName, Namespace: instance.Namespace}, existing); err == nil {
-			log.FromContext(ctx).Info("Cleaning up orphaned Vertex ADC ConfigMap")
-			return r.Delete(ctx, existing)
-		}
-		return nil
-	}
-
-	logger := log.FromContext(ctx)
-
-	cm := &corev1.ConfigMap{}
-	cm.SetName(configMapName)
-	cm.SetNamespace(instance.Namespace)
-	cm.SetGroupVersionKind(corev1.SchemeGroupVersion.WithKind("ConfigMap"))
-	cm.Data = map[string]string{
-		"adc.json": `{"type":"authorized_user","client_id":"stub.apps.googleusercontent.com","client_secret":"stub","refresh_token":"proxy-managed-token"}`,
-	}
-
-	if err := controllerutil.SetControllerReference(instance, cm, r.Scheme); err != nil {
-		return fmt.Errorf("failed to set controller reference on vertex ADC config: %w", err)
-	}
-
-	if err := r.Patch(ctx, cm, client.Apply, &client.PatchOptions{
-		FieldManager: "claw-operator",
-		Force:        ptr.To(true),
-	}); err != nil {
-		return fmt.Errorf("failed to apply vertex ADC config: %w", err)
-	}
-
-	logger.Info("Applied Vertex ADC ConfigMap")
-	return nil
-}
-
 // configureProxyImage overrides the proxy Deployment's container image.
 // If image is empty, the embedded default is preserved.
 func configureProxyImage(objects []*unstructured.Unstructured, instance *clawv1alpha1.Claw, image string) error {
@@ -413,46 +266,6 @@ func configureProxyImage(objects []*unstructured.Unstructured, instance *clawv1a
 		return fmt.Errorf("container %q not found in proxy deployment", ClawProxyContainerName)
 	}
 	return fmt.Errorf("claw-proxy deployment not found in manifests")
-}
-
-// configureImagePullPolicy overrides imagePullPolicy on all containers in all
-// Deployment objects. If policy is empty, the embedded defaults are preserved.
-func configureImagePullPolicy(objects []*unstructured.Unstructured, policy string) error {
-	if policy == "" {
-		return nil
-	}
-
-	for _, obj := range objects {
-		if obj.GetKind() != DeploymentKind {
-			continue
-		}
-
-		for _, path := range [][]string{
-			{"spec", "template", "spec", "containers"},
-			{"spec", "template", "spec", "initContainers"},
-		} {
-			containers, found, err := unstructured.NestedSlice(obj.Object, path...)
-			if err != nil {
-				return fmt.Errorf("failed to get %s from %s: %w", path[len(path)-1], obj.GetName(), err)
-			}
-			if !found {
-				continue
-			}
-
-			for i, c := range containers {
-				cm, ok := c.(map[string]any)
-				if !ok {
-					continue
-				}
-				cm["imagePullPolicy"] = policy
-				containers[i] = cm
-			}
-			if err := unstructured.SetNestedSlice(obj.Object, containers, path...); err != nil {
-				return fmt.Errorf("failed to set %s in %s: %w", path[len(path)-1], obj.GetName(), err)
-			}
-		}
-	}
-	return nil
 }
 
 // configureProxyForCredentials adds credential env vars and volume mounts to the
@@ -578,295 +391,6 @@ func configureProxyForCredentials(objects []*unstructured.Unstructured, instance
 	return fmt.Errorf("claw-proxy deployment not found in manifests")
 }
 
-// configureClawDeploymentForVertex adds Vertex AI environment variables and the stub
-// ADC volume mount to the claw (gateway) deployment when any credential uses the
-// native Vertex SDK (GCP + non-Google provider). The stub ADC allows google-auth-library
-// to bootstrap token refresh, which the MITM proxy intercepts with real credentials.
-func configureClawDeploymentForVertex(objects []*unstructured.Unstructured, credentials []resolvedCredential, instanceName string) error {
-	var vertexCreds []clawv1alpha1.CredentialSpec
-	for _, rc := range credentials {
-		if usesVertexSDK(rc.CredentialSpec) {
-			vertexCreds = append(vertexCreds, rc.CredentialSpec)
-		}
-	}
-	if len(vertexCreds) == 0 {
-		return nil
-	}
-
-	gatewayName := getClawDeploymentName(instanceName)
-	for _, obj := range objects {
-		if obj.GetKind() != DeploymentKind || obj.GetName() != gatewayName {
-			continue
-		}
-
-		containers, found, err := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "containers")
-		if err != nil {
-			return fmt.Errorf("failed to get containers from claw deployment: %w", err)
-		}
-		if !found {
-			return fmt.Errorf("containers field not found in claw deployment")
-		}
-
-		containerIdx := -1
-		var container map[string]any
-		for i, c := range containers {
-			cm, ok := c.(map[string]any)
-			if !ok {
-				continue
-			}
-			if name, _, _ := unstructured.NestedString(cm, "name"); name == ClawGatewayContainerName {
-				containerIdx = i
-				container = cm
-				break
-			}
-		}
-		if containerIdx < 0 {
-			return fmt.Errorf("container %q not found in claw deployment", ClawGatewayContainerName)
-		}
-
-		envVars, _, _ := unstructured.NestedSlice(container, "env")
-		volumeMounts, _, _ := unstructured.NestedSlice(container, "volumeMounts")
-		volumes, _, _ := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "volumes")
-
-		envVars = append(envVars, map[string]any{
-			"name":  "GOOGLE_APPLICATION_CREDENTIALS",
-			"value": "/etc/vertex-adc/adc.json",
-		})
-
-		for _, cred := range vertexCreds {
-			if cred.Provider == "anthropic" && cred.GCP != nil {
-				envVars = append(envVars, map[string]any{
-					"name":  "ANTHROPIC_VERTEX_PROJECT_ID",
-					"value": cred.GCP.Project,
-				})
-				break
-			}
-		}
-
-		volumeMounts = append(volumeMounts, map[string]any{
-			"name":      "vertex-adc",
-			"mountPath": "/etc/vertex-adc",
-			"readOnly":  true,
-		})
-		volumes = append(volumes, map[string]any{
-			"name": "vertex-adc",
-			"configMap": map[string]any{
-				"name": getVertexADCConfigMapName(instanceName),
-			},
-		})
-
-		if err := unstructured.SetNestedSlice(container, envVars, "env"); err != nil {
-			return fmt.Errorf("failed to set env vars on claw deployment: %w", err)
-		}
-		if err := unstructured.SetNestedSlice(container, volumeMounts, "volumeMounts"); err != nil {
-			return fmt.Errorf("failed to set volume mounts on claw deployment: %w", err)
-		}
-		containers[containerIdx] = container
-		if err := unstructured.SetNestedSlice(obj.Object, containers, "spec", "template", "spec", "containers"); err != nil {
-			return fmt.Errorf("failed to set containers on claw deployment: %w", err)
-		}
-		if err := unstructured.SetNestedSlice(obj.Object, volumes, "spec", "template", "spec", "volumes"); err != nil {
-			return fmt.Errorf("failed to set volumes on claw deployment: %w", err)
-		}
-
-		return nil
-	}
-	return fmt.Errorf("claw deployment not found in manifests")
-}
-
-// configureClawDeploymentForKubernetes mounts the sanitized kubeconfig ConfigMap,
-// sets KUBECONFIG and PATH env vars, and adds an init container that copies kubectl
-// into a shared volume on the claw (gateway) deployment when a kubernetes credential is present.
-func configureClawDeploymentForKubernetes(objects []*unstructured.Unstructured, credentials []resolvedCredential, kubectlImage, instanceName string) error {
-	if !hasKubernetesCredentials(credentials) {
-		return nil
-	}
-
-	gatewayName := getClawDeploymentName(instanceName)
-	for _, obj := range objects {
-		if obj.GetKind() != DeploymentKind || obj.GetName() != gatewayName {
-			continue
-		}
-
-		containers, found, err := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "containers")
-		if err != nil {
-			return fmt.Errorf("failed to get containers from claw deployment: %w", err)
-		}
-		if !found {
-			return fmt.Errorf("containers field not found in claw deployment")
-		}
-
-		containerIdx := -1
-		var container map[string]any
-		for i, c := range containers {
-			cm, ok := c.(map[string]any)
-			if !ok {
-				continue
-			}
-			if name, _, _ := unstructured.NestedString(cm, "name"); name == ClawGatewayContainerName {
-				containerIdx = i
-				container = cm
-				break
-			}
-		}
-		if containerIdx < 0 {
-			return fmt.Errorf("container %q not found in claw deployment", ClawGatewayContainerName)
-		}
-
-		envVars, _, _ := unstructured.NestedSlice(container, "env")
-		volumeMounts, _, _ := unstructured.NestedSlice(container, "volumeMounts")
-		volumes, _, _ := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "volumes")
-		initContainers, _, _ := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "initContainers")
-
-		envVars = append(envVars,
-			map[string]any{
-				"name":  "KUBECONFIG",
-				"value": "/etc/kube/config",
-			},
-			map[string]any{
-				"name":  "PATH",
-				"value": "/opt/kube-tools:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-			},
-		)
-
-		volumeMounts = append(volumeMounts,
-			map[string]any{
-				"name":      "kube-config",
-				"mountPath": "/etc/kube",
-				"readOnly":  true,
-			},
-			map[string]any{
-				"name":      "kubectl-bin",
-				"mountPath": "/opt/kube-tools",
-				"readOnly":  true,
-			},
-		)
-
-		volumes = append(volumes,
-			map[string]any{
-				"name": "kube-config",
-				"configMap": map[string]any{
-					"name": getKubeConfigMapName(instanceName),
-				},
-			},
-			map[string]any{
-				"name":     "kubectl-bin",
-				"emptyDir": map[string]any{},
-			},
-		)
-
-		initContainers = append(initContainers, map[string]any{
-			"name":            "init-kubectl",
-			"image":           kubectlImage,
-			"imagePullPolicy": "IfNotPresent",
-			"command":         []any{"sh", "-c", "cp /usr/bin/oc /usr/bin/kubectl /tools/"},
-			"securityContext": map[string]any{
-				"runAsNonRoot":             true,
-				"allowPrivilegeEscalation": false,
-				"readOnlyRootFilesystem":   true,
-				"capabilities":             map[string]any{"drop": []any{"ALL"}},
-			},
-			"resources": map[string]any{
-				"requests": map[string]any{"memory": "32Mi", "cpu": "50m"},
-				"limits":   map[string]any{"memory": "64Mi", "cpu": "100m"},
-			},
-			"volumeMounts": []any{
-				map[string]any{
-					"name":      "kubectl-bin",
-					"mountPath": "/tools",
-				},
-			},
-		})
-
-		if err := unstructured.SetNestedSlice(container, envVars, "env"); err != nil {
-			return fmt.Errorf("failed to set env vars on claw deployment: %w", err)
-		}
-		if err := unstructured.SetNestedSlice(container, volumeMounts, "volumeMounts"); err != nil {
-			return fmt.Errorf("failed to set volume mounts on claw deployment: %w", err)
-		}
-		containers[containerIdx] = container
-		if err := unstructured.SetNestedSlice(obj.Object, containers, "spec", "template", "spec", "containers"); err != nil {
-			return fmt.Errorf("failed to set containers on claw deployment: %w", err)
-		}
-		if err := unstructured.SetNestedSlice(obj.Object, volumes, "spec", "template", "spec", "volumes"); err != nil {
-			return fmt.Errorf("failed to set volumes on claw deployment: %w", err)
-		}
-		if err := unstructured.SetNestedSlice(obj.Object, initContainers, "spec", "template", "spec", "initContainers"); err != nil {
-			return fmt.Errorf("failed to set init containers on claw deployment: %w", err)
-		}
-
-		return nil
-	}
-	return fmt.Errorf("claw deployment not found in manifests")
-}
-
-// configureClawDeploymentConfigMode sets the CLAW_CONFIG_MODE env var on the
-// init-config init container in the claw (gateway) deployment. This controls
-// whether the merge script deep-merges operator config into the existing user
-// config ("merge") or fully overwrites it ("overwrite").
-func configureClawDeploymentConfigMode(objects []*unstructured.Unstructured, instance *clawv1alpha1.Claw) error {
-	mode := string(instance.Spec.ConfigMode)
-	if mode == "" {
-		mode = string(clawv1alpha1.ConfigModeMerge)
-	}
-
-	gatewayName := getClawDeploymentName(instance.Name)
-	for _, obj := range objects {
-		if obj.GetKind() != DeploymentKind || obj.GetName() != gatewayName {
-			continue
-		}
-
-		initContainers, found, err := unstructured.NestedSlice(obj.Object, "spec", "template", "spec", "initContainers")
-		if err != nil {
-			return fmt.Errorf("failed to get init containers from claw deployment: %w", err)
-		}
-		if !found {
-			return fmt.Errorf("initContainers field not found in claw deployment")
-		}
-
-		for i, ic := range initContainers {
-			container, ok := ic.(map[string]any)
-			if !ok {
-				continue
-			}
-			name, _, _ := unstructured.NestedString(container, "name")
-			if name != ClawInitConfigContainerName {
-				continue
-			}
-
-			envVars, _, _ := unstructured.NestedSlice(container, "env")
-
-			updated := false
-			for j, e := range envVars {
-				env, ok := e.(map[string]any)
-				if !ok {
-					continue
-				}
-				if env["name"] == ClawConfigModeEnvVar {
-					env["value"] = mode
-					envVars[j] = env
-					updated = true
-					break
-				}
-			}
-			if !updated {
-				envVars = append(envVars, map[string]any{
-					"name":  ClawConfigModeEnvVar,
-					"value": mode,
-				})
-			}
-
-			if err := unstructured.SetNestedSlice(container, envVars, "env"); err != nil {
-				return fmt.Errorf("failed to set env vars on init-config container: %w", err)
-			}
-			initContainers[i] = container
-			return unstructured.SetNestedSlice(obj.Object, initContainers, "spec", "template", "spec", "initContainers")
-		}
-		return fmt.Errorf("container %q not found in claw deployment", ClawInitConfigContainerName)
-	}
-	return fmt.Errorf("claw deployment not found in manifests")
-}
-
 // stampProxyConfigHash adds a hash annotation to the proxy pod template to trigger
 // rollouts when the proxy config changes.
 func stampProxyConfigHash(objects []*unstructured.Unstructured, instance *clawv1alpha1.Claw, hash string) error {
@@ -887,62 +411,6 @@ func stampProxyConfigHash(objects []*unstructured.Unstructured, instance *clawv1
 		return nil
 	}
 	return fmt.Errorf("claw-proxy deployment not found for config hash stamping")
-}
-
-// stampGatewayConfigHash computes a SHA-256 hash of the gateway ConfigMap data and
-// stamps it on the gateway pod template. This triggers a rollout when operator.json
-// or other operator-managed config files change (e.g., after an operator upgrade).
-func stampGatewayConfigHash(objects []*unstructured.Unstructured, instanceName string) error {
-	configMapName := getConfigMapName(instanceName)
-	var configData map[string]any
-	for _, obj := range objects {
-		if obj.GetKind() == ConfigMapKind && obj.GetName() == configMapName {
-			var found bool
-			var err error
-			configData, found, err = unstructured.NestedMap(obj.Object, "data")
-			if err != nil {
-				return fmt.Errorf("failed to extract data from ConfigMap %s: %w", configMapName, err)
-			}
-			if !found {
-				return fmt.Errorf("data not found in ConfigMap %s", configMapName)
-			}
-			break
-		}
-	}
-	if configData == nil {
-		return fmt.Errorf("ConfigMap %s not found in manifests", configMapName)
-	}
-
-	keys := make([]string, 0, len(configData))
-	for k := range configData {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	h := sha256.New()
-	for _, k := range keys {
-		_, _ = fmt.Fprintf(h, "%s=%v\n", k, configData[k])
-	}
-	hash := fmt.Sprintf("%x", h.Sum(nil))
-
-	gatewayName := getClawDeploymentName(instanceName)
-	for _, obj := range objects {
-		if obj.GetKind() != DeploymentKind || obj.GetName() != gatewayName {
-			continue
-		}
-
-		annotations, _, _ := unstructured.NestedStringMap(obj.Object, "spec", "template", "metadata", "annotations")
-		if annotations == nil {
-			annotations = make(map[string]string)
-		}
-		annotations[clawv1alpha1.AnnotationKeyGatewayConfigHash] = hash
-
-		if err := unstructured.SetNestedStringMap(obj.Object, annotations, "spec", "template", "metadata", "annotations"); err != nil {
-			return fmt.Errorf("failed to set gateway config hash annotation: %w", err)
-		}
-		return nil
-	}
-	return fmt.Errorf("gateway deployment %s not found for config hash stamping", gatewayName)
 }
 
 // stampSecretVersionAnnotation fetches each credential's referenced Secret and stamps
